practice-app/backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.2.1
docutils==0.19
et-xmlfile==1.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
xlrd==2.0.1
requests==2.30.0
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==3.14.0
MarkupSafe==2.1.2
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0 
itypes==1.2.0 
jinja2==3.1.2
openapi-codec==1.3.2 
simplejson==3.19.1
pyyaml==6.0
uritemplate==4.1.1